requirements.txt
# 成語動畫故事生成系統依賴

# 核心依賴
google-generativeai>=0.3.0
Pillow>=10.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

# 視頻處理
moviepy>=1.0.3

# 數據處理
numpy>=1.24.0

# 可選依賴（用於增強功能）
# opencv-python>=4.8.0  # 用於更高級的圖像處理
# scipy>=1.11.0        # 用於數值計算

# 開發依賴
# black>=23.0.0        # 代碼格式化
# flake8>=6.0.0        # 語法檢查
# mypy>=1.5.0          # 類型檢查
